requirements.txt
streamlit
smolagents
litellm
python-dotenv